try to workaround malformed api
sometimes, sometimes the attachments field is wrapped by another layer of quotation marks out of thin air.

and I guess server also just freaked out and throws 503... when requesting full post...

the information is still available in "stub post" from the single case reported, so a retry is possible.
but it's rare so I probably just gonna leave it...
a manual response is needed if you still want the remains to be salvaged.
or maybe you don't care about the extras that requires a full post json.